:recycle: a little delete callback refactoring
