+++ src/main/kotlin/com/github/pool_party/pull_party_bot/commands/Handlers.kt
@@ -16,6 +16,7 @@ import com.github.pool_party.pull_party_bot.commands.handlers.PartyCommand
 import com.github.pool_party.pull_party_bot.commands.handlers.RemoveCommand
 import com.github.pool_party.pull_party_bot.commands.handlers.RudeCommand
 import com.github.pool_party.pull_party_bot.commands.handlers.StartCommand
+import com.github.pool_party.pull_party_bot.commands.handlers.callback.DeleteNodeSuggestionCallback
 import com.github.pool_party.pull_party_bot.commands.handlers.callback.DeleteSuggestionCallback
 import com.github.pool_party.pull_party_bot.commands.handlers.callback.PingCallback
 import com.github.pool_party.pull_party_bot.database.dao.ChatDaoImpl
@@ -27,7 +28,11 @@ fun Bot.initHandlers() {
 
     val chatDaoImpl = ChatDaoImpl()
 
-    val callbacks = listOf(DeleteSuggestionCallback(partyDaoImpl), PingCallback(partyDaoImpl))
+    val callbacks = listOf(
+        DeleteNodeSuggestionCallback(partyDaoImpl),
+        DeleteSuggestionCallback(partyDaoImpl),
+        PingCallback(partyDaoImpl)
+    )
 
     val everyMessageInteractions = listOf(MigrationHandler(chatDaoImpl), ImplicitPartyHandler(partyDaoImpl))
